extract www-authenticate and docker-reference as separate crates
