+++ containrs/containrs/Cargo.toml
@@ -7,14 +7,14 @@ edition = "2018"
 [dependencies]
 oci-image = { path = "../oci-image" }
 oci-distribution = { path = "../oci-distribution" }
+www-authenticate = { path = "../www-authenticate" }
+docker-reference = { path = "../docker-reference" }
 
 async-trait = "0.1"
 failure = "0.1"
 headers = "0.2.3"
 hyper = "0.13.0-alpha.1"
 log = "0.4"
-pest = "2.1"
-pest_derive = "2.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_urlencoded = "0.5"
